event_booking/requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.13.2
